requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.0
django-redis==5.2.0
djangorestframework==3.13.1
psycopg2==2.9.2
pytz==2021.3
redis==4.0.2
sqlparse==0.4.2
tzdata==2021.5
wrapt==1.13.3
